feat: add url tag to console
This change does make commands more interactive and more easy to
navigate.
Adding a help command to get a first message if console opens up.
Improve presentation of different commands to be easier to understand.
Add images to readme to make it more attractive.
Extend readme to be more helpful.